// client/src/Api.ts
export interface Diseases {
  /**
   * Note:
   * This is a Primary Key.<pk/>
   * @format integer
   */
  id?: number;
  /**
   * @format character varying
   * @maxLength 255
   */
  name: string;
}

export interface Diagnoses {
  /**
   * Note:
   * This is a Primary Key.<pk/>
   * @format integer
   */
  id?: number;
  /**
   * Note:
   * This is a Foreign Key to `patients.id`.<fk table='patients' column='id'/>
   * @format integer
   */
  patient_id: number;
  /**
   * Note:
   * This is a Foreign Key to `diseases.id`.<fk table='diseases' column='id'/>
   * @format integer
   */
  disease_id: number;
  /**
   * @format timestamp with time zone
   * @default "CURRENT_TIMESTAMP"
   */
  diagnosis_date?: string;
}

export interface Patients {
    diagnoses: any;
    updated_at: string;
    created_at: string;
    age: string;
  /**
   * Note:
   * This is a Primary Key.<pk/>
   * @format integer
   */
  id?: number;
  /**
   * @format character varying
   * @maxLength 255
   */
  name: string;
}